requirements.txt
crispy-tailwind==0.5.0
Django==3.2.25
django-environ==0.11.2
djangorestframework==3.15.1
Pillow==9.5.0
psycopg2==2.9.9
